et_nVtx attempt 3

+++ Plotting/TL1Plots.h
@@ -15,7 +15,7 @@ class TL1Plots
         ~TL1Plots();
 
         virtual void InitPlots() = 0;
-        virtual void Fill(const double & xVal, const double & yVal, const int & pu=0) = 0;
+        virtual void Fill(const double & xVal, const double & yVal, const double & pu=0) = 0;
         virtual void DrawPlots() = 0;
 
         virtual void SetSample(const std::string & sampleName, const std::string & sampleTitle);
